update item url handling, refactor item_tags
