+++ apps/contrib/templates/meinberlin_contrib/item_detail.html
@@ -1,5 +1,5 @@
 {% extends "base.html" %}
-{% load i18n module_tags rules react_comments react_ratings wagtailcore_tags contrib_tags%}
+{% load i18n module_tags rules react_comments react_ratings wagtailcore_tags item_tags %}
 
 {% block title %}{{object.name}}{% endblock %}
 {% block content %}
@@ -33,10 +33,12 @@
             </button>
             <ul class="dropdown-menu" aria-labelledby="idea-{{object.pk}}-actions">
                 <li>
-                    <a class="dropdown-item" href="{% url 'ideas:idea-update' slug=object.slug %}">{% trans 'Edit' %}</a>
+                    {% get_item_change_url object as change_url %}
+                    <a class="dropdown-item" href="{{ change_url }}">{% trans 'Edit' %}</a>
                 </li>
                 <li>
-                    <a class="dropdown-item" href="{% url 'ideas:idea-delete' slug=object.slug %}">{% trans 'Delete' %}</a>
+                    {% get_item_delete_url object as delete_url %}
+                    <a class="dropdown-item" href="{{ delete_url }}">{% trans 'Delete' %}</a>
                 </li>
                 {% block dropdown_items %}{% endblock %}
             </ul>
